app/backend/reference_doc/requirements.txt
nltk==3.8.1
sentence-transformers==3.0.1
torch==2.3.1
transformers==4.42.4
fastapi[standard]
minio==7.2.13
PyPDF2==3.0.1
beautifulsoup4==4.12.3
html2text
redis==5.2.1
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-jaeger-thrift